+++ packages/runtime-core/src/index.ts
@@ -147,18 +147,19 @@ declare module '@vue/reactivity' {
 }
 
 export {
+  Ref,
+  ToRef,
+  ToRefs,
   ReactiveEffectOptions,
   DebuggerEvent,
   DebuggerOptions,
   TrackOpTypes,
   TriggerOpTypes,
-  Ref,
   ComputedRef,
   WritableComputedRef,
   UnwrapRef,
   ShallowUnwrapRef,
   WritableComputedOptions,
-  ToRefs,
   DeepReadonly
 } from '@vue/reactivity'
 export {