+++ packages/sfc-playground/src/transform.ts
@@ -233,7 +233,7 @@ async function doCompileScript(
       }
 
       return [code, compiledScript.bindings]
-    } catch (e) {
+    } catch (e: any) {
       store.errors = [e.stack.split('\n').slice(0, 12).join('\n')]
       return
     }